+++ multiqc/modules/parse_biosciences/__init__.py
@@ -1,0 +1,3 @@
+from .parse_biosciences import MultiqcModule
+
+__all__ = ["MultiqcModule"]